fix: platform-specific file lock, go 1.26.1, goreleaser deprecated config
  so the codebase compiles and tests pass on Windows
  build tag since the test exercises Unix-only syscall.Flock directly
- Bump go directive to 1.26.1 (fixes 5 stdlib CVEs: GO-2026-4599..4603)
- Run go mod tidy: removes unused golang.org/x/text indirect dep
- Fix .goreleaser.yaml deprecated fields:
  archives.builds -> archives.ids
  format_overrides.format -> format_overrides.formats
