+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/bmf-san/gohan
 
-go 1.26.0
+go 1.26.1
 
 require gopkg.in/yaml.v3 v3.0.1
 
@@ -14,5 +14,4 @@ require (
 require (
 	github.com/dlclark/regexp2 v1.11.5 // indirect
 	golang.org/x/sys v0.13.0 // indirect
-	golang.org/x/text v0.34.0 // indirect
 )
